requirements.txt
asgiref==3.6.0
Django==4.2
django-environ==0.10.0
flake8==6.0.0
sqlparse==0.4.4
mock==5.0.2
coverage==7.2.5
mccabe==0.7.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
drf-yasg==1.21.5

